Chore: move background tracker for thread responses (#510)
* Chore: move background tracker for thread responses

* use abstract class

* rm comment
